moved some services info to home
